Fix exfat_read_sector() and exfat_write_sector() for 64-bit LBA
The functions could not be used to do I/O on devices larger than 2^32
512-byte logical blocks(> 2TB devices). Change the LBA type to 64-bit
type.

+++ include/libexfat.h
@@ -226,10 +226,8 @@ int exfat_set_volume_label(struct exfat *exfat, char *label_input);
 int __exfat_set_volume_guid(struct exfat_dentry *dentry, const char *guid);
 int exfat_read_volume_guid(struct exfat *exfat);
 int exfat_set_volume_guid(struct exfat *exfat, const char *guid);
-int exfat_read_sector(struct exfat_blk_dev *bd, void *buf,
-		unsigned int sec_off);
-int exfat_write_sector(struct exfat_blk_dev *bd, void *buf,
-		unsigned int sec_off);
+int exfat_read_sector(struct exfat_blk_dev *bd, void *buf, unsigned long long sec_off);
+int exfat_write_sector(struct exfat_blk_dev *bd, void *buf, unsigned long long sec_off);
 int exfat_write_checksum_sector(struct exfat_blk_dev *bd,
 	struct exfat_user_input *ui, unsigned int checksum,
 	bool is_backup);
